pyproject.toml
[project]
name = "plantao-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.2",
    "dateparser>=1.3.0",
    "fastapi>=0.129.0",
    "openai>=1.30.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pydantic-settings>=2.13.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.30.0",
    "greenlet>=3.0.0",
    "uvicorn[standard]>=0.41.0",
    "openinference-instrumentation-agno>=0.1.28",
]
